Move remaining last-saved tests to the feature
