+++ test/client/feature/last-saved.spec.js
@@ -11,6 +11,7 @@
  */
 
 import connection from '../../../public/js/src/module/connection';
+import encryptor from '../../../public/js/src/module/encryptor';
 import formCache from '../../../public/js/src/module/form-cache';
 import records from '../../../public/js/src/module/records-queue';
 import settings from '../../../public/js/src/module/settings';
@@ -154,6 +155,287 @@ describe( 'Support for jr://instance/last-saved endpoint', () => {
             store.record.removeAll(),
             store.survey.removeAll(),
         ] ).then( () => done(), done );
+    } );
+
+    describe( 'surveys', () => {
+        /**
+         * @param {Partial<GetFormPartsStubResult>} updates
+         */
+        const updateSurvey = ( updates ) => {
+            // Ensure `_updateCache` receives a new hash indicating it should perform an update
+            sandbox.stub( connection, 'getFormPartsHash' ).callsFake( () => {
+                return Promise.resolve( updates.hash );
+            } );
+
+            let updatePromise = new Promise( resolve => {
+                setTimeout( resolve, formCache.CACHE_UPDATE_INITIAL_DELAY + 1 );
+            } );
+
+            const originalStoreUpdate = store.survey.update.bind( store.survey );
+
+            sandbox.stub( store.survey, 'update' ).callsFake( update => {
+                return originalStoreUpdate( update ).then( result => {
+                    if ( update.model === updates.model ) {
+                        timers.tick( 1 );
+                    }
+
+                    return result;
+                } );
+            } );
+
+            timers.tick( formCache.CACHE_UPDATE_INITIAL_DELAY );
+
+            getFormPartsStubResult = Object.assign( {}, getFormPartsStubResult, updates );
+
+            // Wait for `_updateCache` to resolve
+            return updatePromise.then( () => formCache.get( survey ) );
+        };
+
+        const url1 = '/path/to/source.png';
+        const form1 = `<form class="or"><img src="${url1}"/></form>`;
+        const defaultInstanceData = '<data id="modelA"><item>initial</item><meta><instanceID/></meta></data>';
+        const model1 = `<model><instance>${defaultInstanceData}</instance><instance id="last-saved" src="jr://instance/last-saved"/></model>`;
+        const hash1 = '12345';
+
+        const parser = new DOMParser();
+
+        /** @type { Survey } */
+        let survey;
+
+        /** @type {SurveyExternalData} */
+        let lastSavedExternalData;
+
+        /** @type {GetFormPartsStubResult} */
+        let getFormPartsStubResult;
+
+        /** @type {SinonFakeTimers} */
+        let timers;
+
+        /** @type {EnketoRecord} */
+        let record;
+
+        beforeEach( done => {
+            enketoId = 'surveyC';
+
+            record = {
+                draft: false,
+                enketoId,
+                instanceId: 'recordA',
+                name: 'name A',
+                xml: '<data id="modelA"><item>initial</item><meta><instanceID/></meta></data>',
+            };
+
+            survey = {
+                openRosaId: 'formC',
+                openRosaServer: 'http://localhost:3000',
+                enketoId,
+                theme: '',
+            };
+
+            sandbox.stub( settings, 'enketoId' ).get( () => enketoId );
+
+            // Prevent calls to `_updateCache` after tests complete/stubs are restored
+            timers = sinon.useFakeTimers();
+
+            lastSavedExternalData = {
+                id: 'last-saved',
+                src: 'jr://instance/last-saved',
+                xml: parser.parseFromString( defaultInstanceData, 'text/xml' ),
+            };
+
+            getFormPartsStubResult = {
+                externalData: [
+                    lastSavedExternalData,
+                ],
+                form: form1,
+                model: model1,
+                hash: hash1
+            };
+
+            sandbox.stub( connection, 'getFormParts' ).callsFake( survey => {
+                console.log( 'getFormParts' );
+
+                return formCache.getLastSavedRecord( survey.enketoId )
+                    .then( lastSavedRecord => {
+                        if ( lastSavedRecord != null ) {
+                            return { lastSavedRecord };
+                        }
+
+                        return {};
+                    } )
+                    .then( lastSavedData => {
+                        console.log( 'is enc 4?', encryptor.isEncryptionEnabled( survey ) );
+                        const formParts = Object.assign( {
+                            enketoId: survey.enketoId,
+                        }, getFormPartsStubResult, lastSavedData );
+
+                        if ( encryptor.isEncryptionEnabled( survey ) ) {
+                            return encryptor.setEncryptionEnabled( formParts );
+                        }
+
+                        return formParts;
+                    } );
+            } );
+
+            store.init().then( done, done );
+        } );
+
+        afterEach( done => {
+            timers.clearTimeout();
+            timers.clearInterval();
+            timers.restore();
+
+            store.survey.removeAll().then( done, done );
+        } );
+
+        it( 'sets the survey\'s last saved record', done => {
+            const originalRecord = Object.assign( {}, record );
+
+            formCache.init( survey )
+                .then( () => {
+                    return formCache.setLastSavedRecord( enketoId, record );
+                } )
+                .then( survey => {
+                    Object.entries( originalRecord ).forEach( ( [ key, value ] ) => {
+                        expect( survey.lastSavedRecord[ key ] ).to.equal( value );
+                    } );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'preserves the last saved record when a form is updated', done => {
+            const originalRecord = Object.assign( {}, record );
+            const update = Object.assign( {}, survey, {
+                hash: '123456',
+                model: `${model1}<!-- updated -->`,
+            } );
+
+            formCache.init( survey )
+                .then( () => {
+                    return formCache.setLastSavedRecord( enketoId, record );
+                } )
+                .then( () => updateSurvey( update ) )
+                .then( survey => {
+                    expect( survey.hash ).to.equal( update.hash );
+                    expect( survey.model ).to.equal( update.model );
+
+                    Object.entries( originalRecord ).forEach( ( [ key, value ] ) => {
+                        expect( survey.lastSavedRecord[ key ] ).to.equal( value );
+                    } );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'updates last-saved externalData when the last saved record is updated', done => {
+            const updatedItemValue = 'populated';
+            const update = Object.assign( {}, record, {
+                xml: `<data id="surveyA"><item>${updatedItemValue}</item><meta><instanceID>uuid:ea3baa91-74b5-4892-af6f-96267f7fe12e</instanceID></meta></data>`,
+            } );
+
+            formCache.init( survey )
+                .then( () => formCache.setLastSavedRecord( enketoId, update ) )
+                .then( () => formCache.get( survey ) )
+                .then( survey => {
+                    expect( Array.isArray( survey.externalData ) ).to.equal( true );
+                    expect( survey.externalData.length ).to.equal( 1 );
+
+                    const data = survey.externalData[0];
+
+                    expect( data.id ).to.equal( lastSavedExternalData.id );
+                    expect( data.src ).to.equal( lastSavedExternalData.src );
+
+                    /** @type {Element} */
+                    const xmlDocument = data.xml.documentElement;
+
+                    const dataItemValue = xmlDocument.querySelector( 'item' ).innerHTML;
+
+                    expect( dataItemValue ).to.equal( updatedItemValue );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'does not set the survey\'s last saved record when encryption is enabled', done => {
+            const encryptedSurvey = encryptor.setEncryptionEnabled( survey );
+
+            const form = { id: 'abc', version: '2', encryptionKey: 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA5s9p+VdyX1ikG8nnoXLCC9hKfivAp/e1sHr3O15UQ+a8CjR/QV29+cO8zjS/KKgXZiOWvX+gDs2+5k9Kn4eQm5KhoZVw5Xla2PZtJESAd7dM9O5QrqVJ5Ukrq+kG/uV0nf6X8dxyIluNeCK1jE55J5trQMWT2SjDcj+OVoTdNGJ1H6FL+Horz2UqkIObW5/elItYF8zUZcO1meCtGwaPHxAxlvODe8JdKs3eMiIo9eTT4WbH1X+7nJ21E/FBd8EmnK/91UGOx2AayNxM0RN7pAcj47a434LzeM+XCnBztd+mtt1PSflF2CFE116ikEgLcXCj4aklfoON9TwDIQSp0wIDAQAB' };
+
+            formCache.init( encryptedSurvey )
+                .then( () => encryptor.encryptRecord( form, record ) )
+                .then( encryptedRecord => {
+                    return formCache.setLastSavedRecord( enketoId, encryptedRecord );
+                } )
+                .then( survey => {
+                    expect( survey.lastSavedRecord ).to.equal( undefined );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'does not set the survey\'s last saved unencrypted draft record when encryption is enabled', done => {
+            encryptor.setEncryptionEnabled( survey );
+
+            record.draft = true;
+
+            formCache.init( survey )
+                .then( () => {
+                    return formCache.setLastSavedRecord( enketoId, record );
+                } )
+                .then( survey => {
+                    expect( survey.lastSavedRecord ).to.equal( undefined );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'does not set the survey\'s last saved record when the model does not populate a last-saved secondary instance', done => {
+            getFormPartsStubResult = Object.assign( {}, getFormPartsStubResult, {
+                externalData: [],
+            } );
+
+            formCache.init( survey )
+                .then( () => {
+                    return formCache.setLastSavedRecord( enketoId, record );
+                } )
+                .then( survey => {
+                    expect( survey.lastSavedRecord ).to.equal( undefined );
+                } )
+                .then( done, done );
+        } );
+
+        it( 'removes the survey\'s last saved record when the model no longer populates a last-saved secondary instance', done => {
+            const update = Object.assign( {}, survey, {
+                hash: '123456',
+                model: `${model1}<!-- updated -->`,
+                externalData: [],
+            } );
+
+            formCache.init( survey )
+                .then( () => {
+                    return formCache.setLastSavedRecord( enketoId, record );
+                } )
+                .then( () => updateSurvey( update ) )
+                .then( survey => {
+                    expect( survey.lastSavedRecord ).to.equal( undefined );
+                } )
+                .then( done, done );
+
+        } );
+
+        it( 'gets the survey\'s last saved record', done => {
+            const originalRecord = Object.assign( {}, record );
+
+            formCache.init( survey )
+                .then( survey => {
+                    return formCache.setLastSavedRecord( survey.enketoId, record );
+                } )
+                .then( survey => {
+                    return formCache.getLastSavedRecord( survey.enketoId );
+                } )
+                .then( lastSavedRecord => {
+                    Object.entries( originalRecord ).forEach( ( [ key, value ] ) => {
+                        expect( lastSavedRecord[ key ] ).to.equal( value );
+                    } );
+                } )
+                .then( done, done );
+        } );
     } );
 
     describe( 'storage for offline mode', () => {
